requirements.txt
# General Requirements
async_class #==0.5.0
httpx==0.26.0
playwright==1.41.2
undetected-playwright-patch>=1.40.0.post1700587210000
pybrowsers==0.5.2
chrome-fingerprints==1.1

# Math and Others
numpy==1.26.4
# scipy==1.11.4
# Image Processing
# Pillow==10.2.0


# Artificial Intelligence
hcaptcha_challenger>=0.10.1.post2
recognizer==1.1
# yolov5==7.0.13
# sentence_transformers #==2.2.2
# easyocr==1.7.1
# opencv-python~=4.9.0.80
setuptools~=69.1.0
loguru==0.7.2